feat: Add nullability to selected property in Cards component
The `Cards` component now includes a `selected` property that can be nullable. This change allows for more flexibility when selecting a page. The `selectPage` method has been updated to handle the nullable `selected` property.

These changes enhance the functionality of the `Cards` component by allowing for unselecting a page if needed.